pyproject.toml
[project]
name = "cli-agent-project2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.65.0",
    "google-generativeai>=0.8.6",
    "gradio>=6.8.0",
    "openai>=2.24.0",
    "python-dotenv>=1.2.1",
]
